requirements.txt
python-telegram-bot
SQLAlchemy
asyncpg
alembic
pydantic
pydantic-settings
python-dotenv
greenlet
